+++ minifica.api/Program.cs
@@ -27,6 +27,7 @@ builder.Services.AddScoped<IStateRepository, StateRepository>();
 builder.Services.AddScoped<ITownRepository, TownRepository>();
 builder.Services.AddScoped<IPartnerRepository, PartnerRepository>();
 builder.Services.AddScoped<ISchoolBillingInformationRepository, SchoolBillingInformationRepository>();
+builder.Services.AddScoped<ISchoolRepository, SchoolRepository>();
 
 builder.Services.AddScoped<ICountryManager, CountryManager>();
 builder.Services.AddScoped<IModuleManager, ModuleManager>();
@@ -38,6 +39,7 @@ builder.Services.AddScoped<IStateManager, StateManager>();
 builder.Services.AddScoped<ITownManager, TownManager>();
 builder.Services.AddScoped<IPartnerManager, PartnerManager>();
 builder.Services.AddScoped<ISchoolBillingInformationManager, SchoolBillingInformationManager>();
+builder.Services.AddScoped<ISchoolManager, SchoolManager>();
 
 var app = builder.Build();